backend/requirements.txt
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
asttokens==2.4.1
atlastk==0.13.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
confuse==2.0.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.2
fastapi-mail==1.4.1
h11==0.14.0
idna==3.7
ipykernel==6.29.4
ipython==8.23.0
iso8601==1.1.0
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
motor==3.4.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.6.3
pypika-tortoise==0.1.6
python-amazon-sp-api==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tenacity==8.2.3
tornado==6.4
tortoise==0.1.1
tortoise-orm==0.20.0
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
